refactor(core): unify accepted note runtime boundaries
