+++ src/basic_memory/repository/accepted_note_search_repository.py
@@ -8,8 +8,8 @@ from datetime import datetime
 from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from basic_memory.indexing.accepted_note_search import AcceptedNoteSearchRow
-from basic_memory.indexing.project_index_maintenance import delete_project_index_vector_rows
+from basic_memory.repository.accepted_note_search_row import AcceptedNoteSearchRow
+from basic_memory.repository.accepted_note_vector_cleanup import delete_project_index_vector_rows
 
 type SearchIndexSqlValue = str | int | datetime | None
 type SearchIndexSqlParams = dict[str, SearchIndexSqlValue]
