+++ src/basic_memory/schemas/project_index.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from pydantic import BaseModel, Field
 
 if TYPE_CHECKING:
-    from basic_memory.index.local_project import LocalProjectIndexObservation
+    from basic_memory.index.project_indexing import ProjectIndexObservation
     from basic_memory.indexing.project_index_coordinator import ProjectIndexCoordinatorResult
 
 
@@ -28,7 +28,7 @@ class ProjectIndexStatusResponse(BaseModel):
     @classmethod
     def from_observation(
         cls,
-        observation: "LocalProjectIndexObservation",
+        observation: "ProjectIndexObservation",
     ) -> "ProjectIndexStatusResponse":
         return cls(
             total_files=observation.total_files,
