+++ src/basic_memory/services/entity_service.py
@@ -1,43 +1,26 @@
 """Service for managing entities in the database."""
 
 from collections.abc import Callable
-from copy import deepcopy
 from dataclasses import dataclass
-from datetime import datetime
 from pathlib import Path
-from typing import Any, List, Optional, Sequence, Tuple, Union
+from typing import List, Optional, Sequence, Tuple, Union
 
-import frontmatter
-import yaml
 from loguru import logger
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from basic_memory import db
 from basic_memory.config import ProjectConfig, BasicMemoryConfig
-from basic_memory.file_utils import (
-    ParseError,
-    has_frontmatter,
-    parse_frontmatter,
-    remove_frontmatter,
-    dump_frontmatter,
-)
+from basic_memory.file_utils import remove_frontmatter
 from basic_memory.markdown import EntityMarkdown
 from basic_memory.markdown.entity_parser import (
     EntityParser,
-    _coerce_to_string,
     normalize_frontmatter_metadata,
 )
-from basic_memory.markdown.utils import entity_model_from_markdown, schema_to_markdown
+from basic_memory.markdown.utils import entity_model_from_markdown
 from basic_memory.models import Entity as EntityModel
 from basic_memory.models import Observation, Relation
 from basic_memory.models.knowledge import Entity
-from basic_memory.repository import (
-    AcceptedObservationWrite,
-    AcceptedRelationWrite,
-    ObservationRepository,
-    RelationRepository,
-)
-from basic_memory.repository.project_repository import ProjectRepository
+from basic_memory.repository import ObservationRepository, RelationRepository
 from basic_memory.repository.entity_repository import EntityRepository
 from basic_memory.runtime.note_move import normalize_note_move_destination_path
 from basic_memory.schemas import Entity as EntitySchema
@@ -55,9 +38,29 @@ from basic_memory.services.exceptions import (
     EntityNotFoundError,
 )
 from basic_memory.services.link_resolver import LinkResolver
+from basic_memory.services.note_preparation import (
+    NotePreparation,
+    NotePreparationDependencies,
+    PreparedEntityFields,
+    PreparedEntityMove,
+    PreparedEntityWrite,
+    _fenced_code_line_flags as _note_fenced_code_line_flags,
+    apply_prepared_entity_fields,
+    apply_edit_operation as apply_note_edit_operation,
+    insert_relative_to_section as insert_note_relative_to_section,
+    replace_section_content as replace_note_section_content,
+)
 from basic_memory.services.search_service import SearchService
-from basic_memory.utils import build_canonical_permalink
-from basic_memory.workspace_context import workspace_slug_for_canonical_permalinks
+
+__all__ = [
+    "EntityService",
+    "PreparedEntityFields",
+    "PreparedEntityMove",
+    "PreparedEntityWrite",
+    "apply_prepared_entity_fields",
+]
+
+_fenced_code_line_flags = _note_fenced_code_line_flags
 
 
 @dataclass(frozen=True)
@@ -67,233 +70,6 @@ class EntityWriteResult:
     entity: EntityModel
     content: str
     search_content: str
-
-
-@dataclass(frozen=True, slots=True)
-class PreparedEntityFields:
-    """Entity row values that mirror one accepted markdown snapshot."""
-
-    title: str
-    note_type: str
-    entity_metadata: dict[str, Any] | None
-    content_type: str
-    permalink: str | None
-    file_path: str
-    created_at: datetime
-    updated_at: datetime
-
-
-@dataclass(frozen=True)
-class PreparedEntityWrite:
-    """Accepted note state before any persistence side effects happen.
-
-    Prepare methods return this object after all note semantics have been resolved,
-    but before any file writes or database mutations occur.
-
-    Attributes:
-        file_path: Canonical note path implied by the request.
-        markdown_content: Full markdown to persist, including frontmatter.
-        search_content: Frontmatter-stripped content for inline FTS indexing.
-        entity_fields: Typed entity row values that mirror the accepted markdown state.
-        entity_markdown: Parsed markdown reused by the local write path to update
-            entities, observations, and relations without reparsing a second time.
-    """
-
-    file_path: Path
-    markdown_content: str
-    search_content: str
-    entity_fields: PreparedEntityFields
-    entity_markdown: EntityMarkdown
-
-    @property
-    def observations(self) -> list[AcceptedObservationWrite]:
-        """Accepted-write observations, mapped from the already-parsed markdown.
-
-        Lets the DB-first accepted-write path persist the graph without a second
-        parse; the field-for-field mapping mirrors the ORM rows
-        ``update_entity_and_observations`` builds for the file-indexing path.
-        """
-        return [
-            AcceptedObservationWrite(
-                content=obs.content,
-                category=obs.category,
-                context=obs.context,
-                tags=obs.tags,
-            )
-            for obs in self.entity_markdown.observations
-        ]
-
-    @property
-    def relations(self) -> list[AcceptedRelationWrite]:
-        """Accepted-write relations, mapped from the already-parsed markdown.
-
-        Targets stay unresolved here. The accepted runner resolves only safe
-        self-links; the forward-reference job links all other targets later.
-        """
-        return [
-            AcceptedRelationWrite(
-                relation_type=rel.type,
-                target_name=rel.target,
-                context=rel.context,
-            )
-            for rel in self.entity_markdown.relations
-        ]
-
-
-@dataclass(frozen=True, slots=True)
-class PreparedEditTitleReconciliation:
-    """Markdown title state after reconciling an edit with the note's body heading."""
-
-    markdown_content: str
-    title: str
-    metadata: dict[str, Any] | None
-
-
-@dataclass(frozen=True, slots=True)
-class PreparedEntityMove:
-    """Accepted markdown state for a DB-first note move."""
-
-    file_path: Path
-    markdown_content: str
-    search_content: str
-    permalink: str | None
-
-
-def _frontmatter_permalink(value: object) -> str | None:
-    """Return an explicit frontmatter permalink only when YAML parsed a real string."""
-    return value if isinstance(value, str) and value else None
-
-
-def reconcile_prepared_edit_title_from_h1(
-    *,
-    original_markdown: str,
-    markdown_content: str,
-    current_title: str | None,
-    prepared_title: str,
-    metadata: dict[str, Any] | None,
-) -> PreparedEditTitleReconciliation:
-    """Keep PATCH title metadata in step when a direct H1 edit changes the note title."""
-    # The indexing package imports services at module load time, so keep this shared helper import
-    # inside the function to avoid a startup cycle while still preserving one H1 parser.
-    from basic_memory.indexing.accepted_note_search import (
-        accepted_search_content_from_markdown,
-        first_markdown_h1,
-    )
-
-    original_h1 = first_markdown_h1(accepted_search_content_from_markdown(original_markdown))
-    prepared_h1 = first_markdown_h1(accepted_search_content_from_markdown(markdown_content))
-
-    if not prepared_h1 or prepared_h1 == prepared_title:
-        return PreparedEditTitleReconciliation(
-            markdown_content=markdown_content,
-            title=prepared_title,
-            metadata=metadata,
-        )
-
-    # Only infer a title edit when the note previously used the H1 as its title.
-    # If frontmatter and H1 intentionally differed, preserve that explicit metadata.
-    if original_h1 != current_title or prepared_title != current_title:
-        return PreparedEditTitleReconciliation(
-            markdown_content=markdown_content,
-            title=prepared_title,
-            metadata=metadata,
-        )
-
-    post = frontmatter.loads(markdown_content)
-    post.metadata["title"] = prepared_h1
-    reconciled_metadata = metadata
-    if reconciled_metadata is not None:
-        reconciled_metadata = {**reconciled_metadata, "title": prepared_h1}
-
-    return PreparedEditTitleReconciliation(
-        markdown_content=dump_frontmatter(post),
-        title=prepared_h1,
-        metadata=reconciled_metadata,
-    )
-
-
-def _markdown_heading_level(line: str) -> int | None:
-    """Return the ATX heading level of a line, or None when it is not a heading.
-
-    After up to three leading spaces, only 1-6 '#' characters followed by
-    whitespace (or nothing) form a heading, per CommonMark. '#hashtag' text and
-    7+ '#' runs are ordinary content, so boundaries never land on tag lines.
-    """
-    indent = len(line) - len(line.lstrip(" "))
-    if indent > 3:
-        return None
-
-    candidate = line[indent:]
-    if not candidate.startswith("#"):
-        return None
-    level = len(candidate) - len(candidate.lstrip("#"))
-    if level > 6:
-        return None
-    rest = candidate[level:]
-    if rest and not rest.startswith((" ", "\t")):
-        return None
-    return level
-
-
-def _fence_marker(line: str) -> tuple[str, int, str] | None:
-    """Return a CommonMark fence marker, length, and suffix for a delimiter candidate."""
-    indent = len(line) - len(line.lstrip(" "))
-    if indent > 3:
-        return None
-
-    candidate = line[indent:]
-    if not candidate or candidate[0] not in ("`", "~"):
-        return None
-
-    marker = candidate[0]
-    marker_length = len(candidate) - len(candidate.lstrip(marker))
-    if marker_length < 3:
-        return None
-    return marker, marker_length, candidate[marker_length:]
-
-
-def _fenced_code_line_flags(lines: list[str]) -> list[bool]:
-    """Mark which lines sit inside (or delimit) CommonMark fenced code blocks.
-
-    Fenced code often contains '# comment' lines that look exactly like markdown
-    headings. Section matching and boundary detection must skip those lines, or a
-    code comment would match or terminate a section. CommonMark permits backtick
-    and tilde fences indented by at most three spaces; four-space-indented markers
-    are literal code and must not hide later real headings.
-    """
-    flags: list[bool] = []
-    open_marker: str | None = None
-    open_length = 0
-    for line in lines:
-        marker = _fence_marker(line)
-
-        if open_marker is None:
-            if marker is None:
-                flags.append(False)
-                continue
-
-            marker_char, marker_length, suffix = marker
-            # Backtick fence info strings cannot contain a backtick. Treat such a
-            # line as ordinary text instead of opening a fence that never closes.
-            if marker_char == "`" and "`" in suffix:
-                flags.append(False)
-                continue
-
-            flags.append(True)
-            open_marker = marker_char
-            open_length = marker_length
-            continue
-
-        # Every line inside a fence, including its closing delimiter, is skipped
-        # by heading detection. A close must use the same marker, be at least as
-        # long as the opener, and contain only trailing whitespace.
-        flags.append(True)
-        if marker is not None:
-            marker_char, marker_length, suffix = marker
-            if marker_char == open_marker and marker_length >= open_length and not suffix.strip():
-                open_marker = None
-                open_length = 0
-    return flags
 
 
 class EntityService(BaseService[EntityModel]):
@@ -320,7 +96,15 @@ class EntityService(BaseService[EntityModel]):
         self.session_maker = session_maker
         self.search_service = search_service
         self.app_config = app_config
-        self._project_permalink: Optional[str] = None
+        self._note_preparation = NotePreparation(
+            NotePreparationDependencies(
+                entity_parser=entity_parser,
+                entity_repository=entity_repository,
+                file_service=file_service,
+                session_maker=session_maker,
+                app_config=app_config,
+            )
+        )
         # Callable that returns the current user ID (cloud user_profile_id UUID as string).
         # Default returns None for local/CLI usage. Cloud overrides this to read from UserContext.
         self.get_user_id: Callable[[], Optional[str]] = lambda: None
@@ -331,33 +115,12 @@ class EntityService(BaseService[EntityModel]):
         skip_check: bool = False,
         session: AsyncSession | None = None,
     ) -> List[str]:
-        """Detect potential file path conflicts for a given file path.
-
-        This checks for entities with similar file paths that might cause conflicts:
-        - Case sensitivity differences (Finance/file.md vs finance/file.md)
-        - Character encoding differences
-        - Hyphen vs space differences
-        - Unicode normalization differences
-
-        Args:
-            file_path: The file path to check for conflicts
-            skip_check: If True, skip the check and return empty list (optimization for bulk operations)
-
-        Returns:
-            List of file paths that might conflict with the given file path
-        """
-        if skip_check:
-            return []
-
-        from basic_memory.utils import detect_potential_file_conflicts
-
-        # Load only file paths. Conflict detection is on the hot write path and
-        # does not need observations or relations.
-        async with db.scoped_session(self.session_maker, session) as active_session:
-            existing_paths = await self.repository.get_all_file_paths(active_session)
-
-        # Use the enhanced conflict detection utility
-        return detect_potential_file_conflicts(file_path, existing_paths)
+        """Delegate file-path conflict detection to the shared preparation capability."""
+        return await self._note_preparation.detect_file_path_conflicts(
+            file_path,
+            skip_check=skip_check,
+            session=session,
+        )
 
     async def resolve_permalink(
         self,
@@ -366,121 +129,12 @@ class EntityService(BaseService[EntityModel]):
         skip_conflict_check: bool = False,
         session: AsyncSession | None = None,
     ) -> str:
-        """Get or generate unique permalink for an entity.
-
-        Priority:
-        1. If markdown has permalink and it's not used by another file -> use as is
-        2. If markdown has permalink but it's used by another file -> make unique
-        3. For existing files, keep current permalink from db
-        4. Generate new unique permalink from file path
-
-        Enhanced to detect and handle character-related conflicts.
-
-        Note: Uses lightweight repository methods that skip eager loading of
-        observations and relations for better performance during bulk operations.
-        """
-        file_path_str = Path(file_path).as_posix()
-
-        # Check for potential file path conflicts before resolving permalink
-        async with db.scoped_session(self.session_maker, session) as active_session:
-            conflicts = await self.detect_file_path_conflicts(
-                file_path_str, skip_check=skip_conflict_check, session=active_session
-            )
-            if conflicts:
-                logger.warning(
-                    f"Detected potential file path conflicts for '{file_path_str}': {conflicts}"
-                )
-
-            # If markdown has explicit permalink, try to validate it
-            if markdown and markdown.frontmatter.permalink:
-                desired_permalink = markdown.frontmatter.permalink
-                # Use lightweight method - we only need to check file_path
-                existing_file_path = await self.repository.get_file_path_for_permalink(
-                    active_session, desired_permalink
-                )
-
-                # If no conflict or it's our own file, use as is
-                if not existing_file_path or existing_file_path == file_path_str:
-                    return desired_permalink
-
-            # For existing files, try to find current permalink
-            # Use lightweight method - we only need the permalink
-            existing_permalink = await self.repository.get_permalink_for_file_path(
-                active_session, file_path_str
-            )
-            if existing_permalink:
-                return existing_permalink
-
-            # New file - generate permalink
-            if markdown and markdown.frontmatter.permalink:
-                desired_permalink = markdown.frontmatter.permalink
-            else:
-                # Trigger: generating a permalink for a new file
-                # Why: canonical permalinks may require project prefix for global addressing
-                # Outcome: include project slug when enabled in config
-                include_project = True
-                if self.app_config:
-                    include_project = self.app_config.permalinks_include_project
-
-                workspace_permalink = workspace_slug_for_canonical_permalinks()
-                project_permalink = None
-                # Trigger: project-prefixed permalinks are enabled, or organization workspace
-                #   context requires a complete workspace/project canonical permalink.
-                # Why: project slug is the stable middle segment for globally addressable links.
-                # Outcome: fetch and cache the project's permalink before building the canonical URL.
-                if include_project or workspace_permalink:
-                    project_permalink = await self._get_project_permalink(active_session)
-
-                desired_permalink = build_canonical_permalink(
-                    project_permalink,
-                    file_path_str,
-                    include_project=include_project,
-                    workspace_permalink=workspace_permalink,
-                )
-
-            # Make unique if needed - enhanced to handle character conflicts
-            # Use lightweight existence check instead of loading full entity
-            permalink = desired_permalink
-            suffix = 1
-            while await self.repository.permalink_exists(active_session, permalink):
-                permalink = f"{desired_permalink}-{suffix}"
-                suffix += 1
-                logger.debug(f"creating unique permalink: {permalink}")
-
-        return permalink
-
-    async def _get_project_permalink(self, session: AsyncSession) -> Optional[str]:
-        """Get and cache the current project's permalink."""
-        if self._project_permalink is not None:
-            return self._project_permalink
-
-        project_id = self.repository.project_id
-        if project_id is None:  # pragma: no cover
-            return None  # pragma: no cover
-
-        project_repository = ProjectRepository()
-        project = await project_repository.get_by_id(session, project_id)
-        if project:
-            self._project_permalink = project.permalink
-        return self._project_permalink
-
-    def _build_frontmatter_markdown(
-        self, title: str, note_type: str, permalink: str
-    ) -> EntityMarkdown:
-        """Build a minimal EntityMarkdown object for permalink resolution."""
-        from basic_memory.markdown.schemas import EntityFrontmatter
-
-        frontmatter_metadata = {
-            "title": title,
-            "type": note_type,
-            "permalink": permalink,
-        }
-        frontmatter_obj = EntityFrontmatter(metadata=frontmatter_metadata)
-        return EntityMarkdown(
-            frontmatter=frontmatter_obj,
-            content="",
-            observations=[],
-            relations=[],
+        """Delegate permalink resolution to the shared preparation capability."""
+        return await self._note_preparation.resolve_permalink(
+            file_path,
+            markdown,
+            skip_conflict_check=skip_conflict_check,
+            session=session,
         )
 
     def _coerce_schema_input(self, schema: EntitySchema | EntityModel) -> EntitySchema:
@@ -525,125 +179,6 @@ class EntityService(BaseService[EntityModel]):
         else:
             source_schema._permalink = prepared.entity_fields.permalink
 
-    def _apply_schema_frontmatter_overrides(self, schema: EntitySchema) -> EntityMarkdown | None:
-        """Apply schema content frontmatter overrides and return permalink resolution metadata."""
-        if not schema.content or not has_frontmatter(schema.content):
-            return None
-
-        # Trigger: callers supply markdown that already contains frontmatter.
-        # Why: user-authored frontmatter is part of accepted note semantics, not a persistence detail.
-        # Outcome: note_type/permalink derivation happens before any write path decides how to persist.
-        content_frontmatter = parse_frontmatter(schema.content)
-
-        if "type" in content_frontmatter:
-            schema.note_type = _coerce_to_string(content_frontmatter["type"])
-
-        if "permalink" not in content_frontmatter:
-            return None
-
-        content_permalink = _frontmatter_permalink(content_frontmatter["permalink"])
-        if content_permalink is None:
-            return None
-
-        return self._build_frontmatter_markdown(
-            schema.title,
-            schema.note_type,
-            content_permalink,
-        )
-
-    async def _resolve_schema_permalink(
-        self,
-        schema: EntitySchema,
-        *,
-        file_path: Path,
-        current_permalink: str | None = None,
-        content_markdown: EntityMarkdown | None = None,
-        skip_conflict_check: bool = False,
-        session: AsyncSession | None = None,
-    ) -> str | None:
-        """Resolve the canonical permalink for a create/update write."""
-        if self.app_config and self.app_config.disable_permalinks:
-            if current_permalink is None:
-                schema._permalink = ""
-                return None
-            schema._permalink = current_permalink
-            return current_permalink
-
-        if current_permalink and not (content_markdown and content_markdown.frontmatter.permalink):
-            schema._permalink = current_permalink
-            return current_permalink
-
-        resolved_permalink = await self.resolve_permalink(
-            file_path,
-            content_markdown,
-            skip_conflict_check=skip_conflict_check,
-            session=session,
-        )
-        schema._permalink = resolved_permalink
-        return resolved_permalink
-
-    def _build_entity_fields(
-        self,
-        *,
-        file_path: Path,
-        content_type: str,
-        permalink: str | None,
-        entity_markdown: EntityMarkdown,
-    ) -> PreparedEntityFields:
-        """Build the entity row data that mirrors accepted markdown state."""
-        if entity_markdown.created is None or entity_markdown.modified is None:  # pragma: no cover
-            raise ValueError("Prepared markdown requires created and modified timestamps")
-
-        normalized_metadata = normalize_frontmatter_metadata(
-            entity_markdown.frontmatter.metadata or {}
-        )
-        entity_metadata = {k: v for k, v in normalized_metadata.items() if v is not None}
-        return PreparedEntityFields(
-            title=entity_markdown.frontmatter.title,
-            note_type=entity_markdown.frontmatter.type,
-            file_path=file_path.as_posix(),
-            content_type=content_type,
-            entity_metadata=entity_metadata or None,
-            permalink=permalink,
-            created_at=entity_markdown.created,
-            updated_at=entity_markdown.modified,
-        )
-
-    async def _build_prepared_write(
-        self,
-        *,
-        file_path: Path,
-        markdown_content: str,
-        content_type: str,
-        permalink: str | None,
-        preserved_created_at: datetime | None = None,
-    ) -> PreparedEntityWrite:
-        """Parse accepted markdown once so all persistence paths share the same state."""
-        # Trigger: both local and cloud-style callers need the exact same accepted markdown.
-        # Why: parsing twice creates opportunities for drift between "what we accepted" and
-        #      "what we indexed/persisted".
-        # Outcome: callers carry one prepared object through file writes, DB writes, and indexing.
-        entity_markdown = await self.entity_parser.parse_markdown_content(
-            file_path=file_path,
-            content=markdown_content,
-            # DB-first updates have no file ctime. Reuse the existing semantic creation
-            # time as that field's fallback so editing a legacy note cannot make it "new".
-            ctime=(preserved_created_at.timestamp() if preserved_created_at is not None else None),
-        )
-        entity_fields = self._build_entity_fields(
-            file_path=file_path,
-            content_type=content_type,
-            permalink=permalink,
-            entity_markdown=entity_markdown,
-        )
-        return PreparedEntityWrite(
-            file_path=file_path,
-            markdown_content=markdown_content,
-            search_content=remove_frontmatter(markdown_content),
-            entity_fields=entity_fields,
-            entity_markdown=entity_markdown,
-        )
-
     async def _read_persisted_write_content(self, file_path: Path) -> tuple[str, str]:
         """Read the stored markdown after write-time formatting has finished."""
         # Trigger: format-on-save or platform-specific text writes can change the stored markdown
@@ -687,35 +222,11 @@ class EntityService(BaseService[EntityModel]):
               acceptance and must perform any external storage conflict handling
               themselves.
         """
-        # Work on a copy so prepare methods are pure from the caller's perspective.
-        # The router/service layer still receives the same accepted result, but we avoid mutating
-        # the original schema instance in surprising ways.
-        schema = schema.model_copy(deep=True)
-        file_path = Path(schema.file_path)
-
-        if check_storage_exists and await self.file_service.exists(file_path):
-            raise EntityAlreadyExistsError(
-                f"file for entity {schema.directory}/{schema.title} already exists: {file_path}"
-            )
-
-        content_markdown = self._apply_schema_frontmatter_overrides(schema)
-        permalink = await self._resolve_schema_permalink(
+        return await self._note_preparation.prepare_create_entity_content(
             schema,
-            file_path=file_path,
-            content_markdown=content_markdown,
+            check_storage_exists=check_storage_exists,
             skip_conflict_check=skip_conflict_check,
             session=session,
-        )
-
-        # Build the final markdown once here. Local mode will write it immediately; cloud mode can
-        # store it in note_content first and materialize later without re-deriving anything.
-        post = await schema_to_markdown(schema)
-        markdown_content = dump_frontmatter(post)
-        return await self._build_prepared_write(
-            file_path=file_path,
-            markdown_content=markdown_content,
-            content_type=schema.content_type,
-            permalink=permalink,
         )
 
     async def prepare_update_entity_content(
@@ -734,66 +245,12 @@ class EntityService(BaseService[EntityModel]):
         preserve unrecognized frontmatter keys from that explicit base content.
         No database rows are mutated here.
         """
-        schema = schema.model_copy(deep=True)
-        file_path = Path(schema.file_path)
-        current_file_path = Path(entity.file_path)
-        existing_metadata: dict[str, Any] = {}
-        if has_frontmatter(existing_content):
-            try:
-                existing_metadata = parse_frontmatter(existing_content)
-            except ParseError:
-                # Trigger: the old note has frontmatter fences but malformed YAML.
-                # Why: a full replacement must be able to repair that note, and malformed
-                #      metadata cannot be merged safely into the replacement.
-                # Outcome: discard only the invalid merge input; the final accepted markdown
-                #          is still parsed and validated below.
-                pass
-
-        content_markdown = self._apply_schema_frontmatter_overrides(schema)
-        # Trigger: a full replacement may also rename the note by changing title or directory.
-        # Why: cloud accepts the final markdown before S3 is updated, so the prepare contract must
-        #      describe the requested destination instead of silently keeping the old path.
-        # Outcome: unchanged paths preserve the current permalink; renamed paths only rotate the
-        #          permalink when move-policy allows it or frontmatter explicitly sets one.
-        update_permalink_on_rename = bool(
-            self.app_config and self.app_config.update_permalinks_on_move
-        )
-        current_permalink = (
-            entity.permalink
-            if file_path.as_posix() == current_file_path.as_posix()
-            or not update_permalink_on_rename
-            else None
-        )
-        resolved_permalink = await self._resolve_schema_permalink(
+        return await self._note_preparation.prepare_update_entity_content(
+            entity,
             schema,
-            file_path=file_path,
-            current_permalink=current_permalink,
-            content_markdown=content_markdown,
+            existing_content,
             skip_conflict_check=skip_conflict_check,
             session=session,
-        )
-
-        post = await schema_to_markdown(schema)
-
-        # Full updates preserve unrecognized frontmatter keys from the existing note.
-        # That keeps Basic Memory's write semantics stable for hand-authored metadata while still
-        # letting the incoming schema replace the fields it explicitly owns.
-        # Existing frontmatter is a merge input, not accepted state. Semantic validation happens
-        # after the incoming metadata has had a chance to repair invalid canonical values.
-        merged_metadata = deepcopy(existing_metadata)
-        merged_metadata.update(post.metadata)
-        merged_metadata["permalink"] = resolved_permalink
-
-        merged_post = frontmatter.Post(post.content)
-        merged_post.metadata.update(merged_metadata)
-
-        markdown_content = dump_frontmatter(merged_post)
-        return await self._build_prepared_write(
-            file_path=file_path,
-            markdown_content=markdown_content,
-            content_type=schema.content_type,
-            permalink=resolved_permalink,
-            preserved_created_at=entity.created_at,
         )
 
     async def prepare_edit_entity_content(
@@ -817,68 +274,17 @@ class EntityService(BaseService[EntityModel]):
         edit base explicit so higher layers can reject stale content instead of
         silently editing whichever storage copy happens to be newest.
         """
-        file_path = Path(entity.file_path)
-        # Edits are intentionally based on explicit caller-supplied content. That makes stale-base
-        # handling visible to the caller instead of quietly reading whatever persistence layer
-        # happens to be newest.
-        markdown_content = self.apply_edit_operation(
+        return await self._note_preparation.prepare_edit_entity_content(
+            entity,
             current_content,
-            operation,
-            content,
-            section,
-            find_text,
-            expected_replacements,
-            replace_subsections,
-        )
-
-        title = entity.title
-        note_type = entity.note_type
-        permalink = entity.permalink
-        metadata = entity.entity_metadata
-
-        if has_frontmatter(markdown_content):
-            content_frontmatter = parse_frontmatter(markdown_content)
-
-            if "title" in content_frontmatter:
-                title = _coerce_to_string(content_frontmatter["title"])
-            if "type" in content_frontmatter:
-                note_type = _coerce_to_string(content_frontmatter["type"])
-
-            if self.app_config and self.app_config.disable_permalinks:
-                permalink = entity.permalink
-            elif "permalink" in content_frontmatter:
-                content_permalink = _frontmatter_permalink(content_frontmatter["permalink"])
-                if content_permalink is not None:
-                    content_markdown = self._build_frontmatter_markdown(
-                        title,
-                        note_type,
-                        content_permalink,
-                    )
-                    permalink = await self.resolve_permalink(
-                        file_path,
-                        content_markdown,
-                        skip_conflict_check=skip_conflict_check,
-                        session=session,
-                    )
-
-            normalized_metadata = normalize_frontmatter_metadata(content_frontmatter or {})
-            metadata = {k: v for k, v in normalized_metadata.items() if v is not None} or None
-
-        title_reconciliation = reconcile_prepared_edit_title_from_h1(
-            original_markdown=current_content,
-            markdown_content=markdown_content,
-            current_title=entity.title,
-            prepared_title=title,
-            metadata=metadata,
-        )
-        markdown_content = title_reconciliation.markdown_content
-
-        return await self._build_prepared_write(
-            file_path=file_path,
-            markdown_content=markdown_content,
-            content_type=entity.content_type,
-            permalink=permalink,
-            preserved_created_at=entity.created_at,
+            operation=operation,
+            content=content,
+            section=section,
+            find_text=find_text,
+            expected_replacements=expected_replacements,
+            replace_subsections=replace_subsections,
+            skip_conflict_check=skip_conflict_check,
+            session=session,
         )
 
     async def prepare_move_entity_content(
@@ -895,31 +301,11 @@ class EntityService(BaseService[EntityModel]):
         The caller supplies the current accepted markdown because cloud DB-first
         moves may need to use note_content rather than a materialized file.
         """
-        # Keep the search helper import lazy for the same package-cycle reason as
-        # reconcile_prepared_edit_title_from_h1.
-        from basic_memory.indexing.accepted_note_search import (
-            accepted_search_content_from_markdown,
-        )
-
-        file_path = Path(normalize_note_move_destination_path(destination_path))
-        markdown_content = current_content
-        permalink = entity.permalink
-        disable_permalinks = bool(self.app_config and self.app_config.disable_permalinks)
-        update_permalinks_on_move = bool(
-            self.app_config and self.app_config.update_permalinks_on_move
-        )
-
-        if not disable_permalinks and (update_permalinks_on_move or entity.permalink is None):
-            permalink = await self.resolve_permalink(file_path, session=session)
-            post = frontmatter.loads(markdown_content)
-            post.metadata["permalink"] = permalink
-            markdown_content = dump_frontmatter(post)
-
-        return PreparedEntityMove(
-            file_path=file_path,
-            markdown_content=markdown_content,
-            search_content=accepted_search_content_from_markdown(markdown_content),
-            permalink=permalink,
+        return await self._note_preparation.prepare_move_entity_content(
+            entity,
+            current_content,
+            destination_path,
+            session=session,
         )
 
     async def verify_move_destination_absent(
@@ -935,16 +321,10 @@ class EntityService(BaseService[EntityModel]):
         a duplicate. A case-only rename or shared storage target (same physical file)
         is allowed. Cloud is DB-first and opts out via verify_storage_absent_on_create.
         """
-        source = Path(source_file_path)
-        destination = Path(normalize_note_move_destination_path(destination_file_path))
-        if (
-            source != destination
-            and await self.file_service.exists(destination)
-            and not self._paths_share_storage_target(source, destination)
-        ):
-            raise EntityAlreadyExistsError(
-                f"file already exists at destination path: {destination.as_posix()}"
-            )
+        await self._note_preparation.verify_move_destination_absent(
+            source_file_path=source_file_path,
+            destination_file_path=destination_file_path,
+        )
 
     async def create_or_update_entity(self, schema: EntitySchema) -> Tuple[EntityModel, bool]:
         """Create new entity or update existing one.
@@ -1424,35 +804,11 @@ class EntityService(BaseService[EntityModel]):
         self, target: str, entity: EntityModel, session: AsyncSession | None = None
     ) -> EntityModel | None:
         """Resolve only self-relations that are safe to identify in deferred mode."""
-        clean_target = target.strip()
-        if clean_target.startswith("[[") and clean_target.endswith("]]"):
-            clean_target = clean_target[2:-2].strip()
-        if "|" in clean_target:
-            clean_target = clean_target.split("|", 1)[0].strip()
-
-        candidates = {entity.file_path}
-        if entity.permalink:
-            candidates.add(entity.permalink)
-        if entity.file_path.endswith(".md"):
-            candidates.add(entity.file_path[:-3])
-
-        if clean_target in candidates:
-            return entity
-
-        if clean_target != entity.title:
-            return None
-
-        # Title-only links are ambiguous because Basic Memory allows duplicate titles.
-        # Collapse them to self only when the title lookup proves this source is the sole candidate;
-        # otherwise leave the relation unresolved so we do not create a wrong permanent edge.
-        async with db.scoped_session(self.session_maker, session) as active_session:
-            title_matches = await self.repository.get_by_title(
-                active_session, clean_target, load_relations=False
-            )
-            if len(title_matches) == 1 and title_matches[0].id == entity.id:
-                return entity
-
-        return None
+        return await self._note_preparation.resolve_deferred_self_relation(
+            target,
+            entity,
+            session=session,
+        )
 
     async def edit_entity(
         self,
@@ -1572,60 +928,15 @@ class EntityService(BaseService[EntityModel]):
         replace_subsections: bool = True,
     ) -> str:
         """Apply the specified edit operation to the current content."""
-
-        if operation == "append":
-            # Ensure proper spacing
-            if current_content and not current_content.endswith("\n"):
-                return current_content + "\n" + content
-            return current_content + content  # pragma: no cover
-
-        elif operation == "prepend":
-            # Handle frontmatter-aware prepending
-            return self._prepend_after_frontmatter(current_content, content)
-
-        elif operation == "find_replace":
-            if not find_text:
-                raise ValueError("find_text is required for find_replace operation")
-            if not find_text.strip():
-                raise ValueError("find_text cannot be empty or whitespace only")
-
-            # Count actual occurrences
-            actual_count = current_content.count(find_text)
-
-            # Validate count matches expected
-            if actual_count != expected_replacements:
-                if actual_count == 0:
-                    raise ValueError(f"Text to replace not found: '{find_text}'")
-                else:
-                    raise ValueError(
-                        f"Expected {expected_replacements} occurrences of '{find_text}', "
-                        f"but found {actual_count}"
-                    )
-
-            return current_content.replace(find_text, content)
-
-        elif operation == "replace_section":
-            if not section:
-                raise ValueError("section is required for replace_section operation")
-            if not section.strip():
-                raise ValueError("section cannot be empty or whitespace only")
-            return self.replace_section_content(
-                current_content,
-                section,
-                content,
-                replace_subsections=replace_subsections,
-            )
-
-        elif operation in ("insert_before_section", "insert_after_section"):
-            if not section:
-                raise ValueError("section is required for insert section operations")
-            if not section.strip():
-                raise ValueError("section cannot be empty or whitespace only")
-            position = "before" if operation == "insert_before_section" else "after"
-            return self.insert_relative_to_section(current_content, section, content, position)
-
-        else:
-            raise ValueError(f"Unsupported operation: {operation}")
+        return apply_note_edit_operation(
+            current_content,
+            operation,
+            content,
+            section,
+            find_text,
+            expected_replacements,
+            replace_subsections,
+        )
 
     def replace_section_content(
         self,
@@ -1665,67 +976,12 @@ class EntityService(BaseService[EntityModel]):
         Raises:
             ValueError: If multiple sections with the same header are found
         """
-        # Normalize the section header (ensure it starts with #)
-        if not section_header.startswith("#"):
-            section_header = "## " + section_header
-
-        # Strip duplicate header from new_content if present (fix for issue #390)
-        # LLMs sometimes include the section header in their content, which would create duplicates
-        new_content_lines = new_content.lstrip().split("\n")
-        if new_content_lines and new_content_lines[0].strip() == section_header.strip():
-            # Remove the duplicate header line
-            new_content = "\n".join(new_content_lines[1:]).lstrip()
-
-        lines = current_content.split("\n")
-        # Fenced code can contain lines identical to the requested header, so section
-        # matching must skip fenced lines or a code sample would trigger the
-        # duplicate-header error (or match as the section itself).
-        fenced = _fenced_code_line_flags(lines)
-
-        # First pass: count matching sections to check for duplicates
-        matching_sections = [
-            i
-            for i, line in enumerate(lines)
-            if not fenced[i] and line.strip() == section_header.strip()
-        ]
-
-        # Handle multiple sections error
-        if len(matching_sections) > 1:
-            raise ValueError(
-                f"Multiple sections found with header '{section_header}'. "
-                f"Section replacement requires unique headers."
-            )
-
-        # If no section found, append it
-        if len(matching_sections) == 0:
-            logger.info(f"Section '{section_header}' not found, appending to end of document")
-            separator = "\n\n" if current_content and not current_content.endswith("\n\n") else ""
-            return current_content + separator + section_header + "\n" + new_content
-
-        # Replace the single matching section. The replaced span always ends at a
-        # boundary computed from the ORIGINAL lines, so headings inside new_content
-        # cannot extend or shorten what gets consumed (issue #1012).
-        section_line_idx = matching_sections[0]
-        target_level = len(section_header) - len(section_header.lstrip("#"))
-
-        end_idx = len(lines)
-        for i in range(section_line_idx + 1, len(lines)):
-            if fenced[i]:
-                continue
-            heading_level = _markdown_heading_level(lines[i])
-            if heading_level is None:
-                continue
-            # Level-aware default: an h2 section owns its h3+ subsections, so only a
-            # heading at the same or higher level ends it. The opt-out stops at any
-            # heading, preserving subsections (pre-#1012 behavior).
-            if not replace_subsections or heading_level <= target_level:
-                end_idx = i
-                break
-
-        result_lines = lines[: section_line_idx + 1]
-        result_lines.append(new_content)
-        result_lines.extend(lines[end_idx:])
-        return "\n".join(result_lines)
+        return replace_note_section_content(
+            current_content,
+            section_header,
+            new_content,
+            replace_subsections=replace_subsections,
+        )
 
     def insert_relative_to_section(
         self,
@@ -1752,80 +1008,12 @@ class EntityService(BaseService[EntityModel]):
         Raises:
             ValueError: If the section header is not found or appears more than once
         """
-        # Normalize the section header (ensure it starts with #)
-        if not section_header.startswith("#"):
-            section_header = "## " + section_header
-
-        lines = current_content.split("\n")
-        # Fenced code can contain a line identical to the requested heading; skip
-        # fenced lines so a code sample never anchors (or duplicates) the section.
-        fenced = _fenced_code_line_flags(lines)
-        matching_indices = [
-            i
-            for i, line in enumerate(lines)
-            if not fenced[i] and line.strip() == section_header.strip()
-        ]
-
-        if len(matching_indices) == 0:
-            raise ValueError(
-                f"Section '{section_header}' not found in document. "
-                f"Use replace_section to create a new section."
-            )
-        if len(matching_indices) > 1:
-            raise ValueError(
-                f"Multiple sections found with header '{section_header}'. "
-                f"Section insertion requires unique headers."
-            )
-
-        idx = matching_indices[0]
-
-        if position == "before":
-            # Insert new content before the section heading
-            before = lines[:idx]
-            after = lines[idx:]
-            # Ensure blank line separation
-            insert_lines = new_content.rstrip("\n").split("\n")
-            if before and before[-1].strip() != "":
-                insert_lines = [""] + insert_lines
-            return "\n".join(before + insert_lines + [""] + after)
-        else:
-            # Insert new content after the section heading line
-            before = lines[: idx + 1]
-            after = lines[idx + 1 :]
-            insert_lines = new_content.rstrip("\n").split("\n")
-            # Ensure blank line separation so inserted text doesn't merge
-            # with existing section content into a single paragraph
-            if after and after[0].strip() != "":
-                insert_lines = insert_lines + [""]
-            return "\n".join(before + insert_lines + after)
-
-    def _prepend_after_frontmatter(self, current_content: str, content: str) -> str:
-        """Prepend content after frontmatter, preserving frontmatter structure."""
-
-        # Trigger: the note starts with frontmatter delimiters.
-        # Why: prepend must preserve the existing YAML block and insert content into the body,
-        #      not silently rewrite malformed metadata into a corrupted accepted note state.
-        # Outcome: valid frontmatter is preserved, and malformed frontmatter fails fast.
-        if has_frontmatter(current_content):
-            # Parse and separate frontmatter from body. Parse errors are intentional caller-visible
-            # failures so prepare_edit_entity_content can reject unsafe accepted writes.
-            frontmatter_data = parse_frontmatter(current_content)
-            body_content = remove_frontmatter(current_content)
-
-            # Prepend content to the body
-            if content and not content.endswith("\n"):
-                new_body = content + "\n" + body_content
-            else:
-                new_body = content + body_content
-
-            # Reconstruct file with frontmatter + prepended body
-            yaml_fm = yaml.dump(frontmatter_data, sort_keys=False, allow_unicode=True)
-            return f"---\n{yaml_fm}---\n\n{new_body.strip()}"
-
-        # No frontmatter means prepend is a plain text edit.
-        if content and not content.endswith("\n"):
-            return content + "\n" + current_content
-        return content + current_content
+        return insert_note_relative_to_section(
+            current_content,
+            section_header,
+            new_content,
+            position,
+        )
 
     async def move_entity(
         self,
