+++ tests/cloud/test_project_deletes.py
@@ -4,7 +4,7 @@ from collections.abc import AsyncGenerator
 
 import pytest
 import pytest_asyncio
-from basic_memory.cloud.project_deletes import (
+from basic_memory.services.project_deletes import (
     ProjectDeleteAcceptanceError,
     ProjectDeleteAcceptanceRequest,
     ProjectDeleteAcceptanceService,
